requirements.txt
collections
datasets
numpy
matplotlib
seaborn
pandas
scikit-learn
joblib
torch
torchvision
pillow
streamlit